+++ locals.tf
@@ -1,0 +1,4 @@
+locals {
+  trusted_principal_arns = var.trusted_principal_arns
+}
+