+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.3] - 2025-11-24
+
+### Added
+- New `trusted_principal_arns` variable to configure which AWS principals can assume the Leapfrog integration IAM role.
+
+### Changed
+- Leapfrog integration role trust policy now derives principals from `trusted_principal_arns` instead of a hardcoded AWS account ID.
+
 ## [1.0.2] - 2025-11-24
 
 ### Fixed